+++ pulp/0.2/label_balance_edges_maxcut.cpp
@@ -59,10 +59,10 @@ void label_balance_edges_maxcut(pulp_graph_t& g, int64_t num_parts, int64_t* par
   double vert_balance, double edge_balance)
 {
   int64_t num_verts = g.n;
-  unsigned num_edges = g.m;
-  unsigned cut_size = 0;
+  uint64_t num_edges = g.m;
+  uint64_t cut_size = 0;
   int64_t* part_sizes = new int64_t[num_parts];
-  unsigned* part_edge_sizes = new unsigned[num_parts];
+  uint64_t* part_edge_sizes = new uint64_t[num_parts];
   int64_t* part_cut_sizes = new int64_t[num_parts];
  
   for (int64_t i = 0; i < num_parts; ++i)
@@ -95,9 +95,9 @@ void label_balance_edges_maxcut(pulp_graph_t& g, int64_t num_parts, int64_t* par
 {
 
   int64_t* part_sizes_thread = new int64_t[num_parts];
-  unsigned* part_edge_sizes_thread = new unsigned[num_parts];
+  uint64_t* part_edge_sizes_thread = new uint64_t[num_parts];
   int64_t* part_cut_sizes_thread = new int64_t[num_parts];
-  unsigned cut_size_thread = 0;
+  uint64_t cut_size_thread = 0;
   for (int64_t i = 0; i < num_parts; ++i) 
     part_sizes_thread[i] = 0;
   for (int64_t i = 0; i < num_parts; ++i) 
@@ -109,12 +109,12 @@ void label_balance_edges_maxcut(pulp_graph_t& g, int64_t num_parts, int64_t* par
   for (int64_t i = 0; i < num_verts; ++i)
   {
     int64_t part = parts[i];
-    unsigned out_degree = out_degree(g, i);
+    uint64_t out_degree = out_degree(g, i);
     ++part_sizes_thread[part];
     part_edge_sizes_thread[part] += out_degree;
 
     int64_t* outs = out_vertices(g, i);
-    for (unsigned j = 0; j < out_degree; ++j)
+    for (uint64_t j = 0; j < out_degree; ++j)
     {
       int64_t out = outs[j];
       int64_t out_part = parts[out];
@@ -218,9 +218,9 @@ while(t < edge_outer_iter)
       for (int64_t p = 0; p < num_parts; ++p)
         part_counts[p] = 0.0;
 
-      unsigned out_degree = out_degree(g, v);
+      uint64_t out_degree = out_degree(g, v);
       int64_t* outs = out_vertices(g, v);
-      for (unsigned j = 0; j < out_degree; ++j)
+      for (uint64_t j = 0; j < out_degree; ++j)
       {
         int64_t out = outs[j];
         int64_t part_out = parts[out];
@@ -285,7 +285,7 @@ while(t < edge_outer_iter)
             thread_queue_size = 0;
           }
         }
-        for (unsigned j = 0; j < out_degree; ++j)
+        for (uint64_t j = 0; j < out_degree; ++j)
         {
           if (!in_queue_next[outs[j]])
           {
@@ -420,9 +420,9 @@ while(t < edge_outer_iter)
         part_counts[p] = 0;
 
       int64_t part = parts[v];
-      unsigned out_degree = out_degree(g, v);
+      uint64_t out_degree = out_degree(g, v);
       int64_t* outs = out_vertices(g, v);
-      for (unsigned j = 0; j < out_degree; ++j)
+      for (uint64_t j = 0; j < out_degree; ++j)
       {
         int64_t out = outs[j];
         int64_t part_out = parts[out];
@@ -487,7 +487,7 @@ while(t < edge_outer_iter)
               thread_queue_size = 0;
             }
           }
-          for (unsigned j = 0; j < out_degree; ++j) 
+          for (uint64_t j = 0; j < out_degree; ++j) 
           {
             if (!in_queue_next[outs[j]])
             {
@@ -623,15 +623,15 @@ void label_balance_edges_maxcut_weighted(
   double vert_balance, double edge_balance)
 {
   int64_t num_verts = g.n;
-  unsigned num_edges = g.m;
-  unsigned cut_size = 0;
+  uint64_t num_edges = g.m;
+  uint64_t cut_size = 0;
 
   bool has_vwgts = (g.vertex_weights != NULL);
   bool has_ewgts = (g.edge_weights != NULL);
   if (!has_vwgts) g.vertex_weights_sum = g.n;
 
   int64_t* part_sizes = new int64_t[num_parts];
-  unsigned* part_edge_sizes = new unsigned[num_parts];
+  uint64_t* part_edge_sizes = new uint64_t[num_parts];
   int64_t* part_cut_sizes = new int64_t[num_parts];
  
   for (int64_t i = 0; i < num_parts; ++i)
@@ -664,9 +664,9 @@ void label_balance_edges_maxcut_weighted(
 {
 
   int64_t* part_sizes_thread = new int64_t[num_parts];
-  unsigned* part_edge_sizes_thread = new unsigned[num_parts];
+  uint64_t* part_edge_sizes_thread = new uint64_t[num_parts];
   int64_t* part_cut_sizes_thread = new int64_t[num_parts];
-  unsigned cut_size_thread = 0;
+  uint64_t cut_size_thread = 0;
   
   for (int64_t i = 0; i < num_parts; ++i) 
     part_sizes_thread[i] = 0;
@@ -679,14 +679,14 @@ void label_balance_edges_maxcut_weighted(
   for (int64_t i = 0; i < num_verts; ++i)
   {
     int64_t part = parts[i];
-    unsigned out_degree = out_degree(g, i);
+    uint64_t out_degree = out_degree(g, i);
     if (has_vwgts) part_sizes_thread[part] += g.vertex_weights[i];
     else ++part_sizes_thread[part];
     part_edge_sizes_thread[part] += out_degree;
 
     int64_t* outs = out_vertices(g, i);
     int64_t* weights = out_weights(g, i);
-    for (unsigned j = 0; j < out_degree; ++j)
+    for (uint64_t j = 0; j < out_degree; ++j)
     {
       int64_t out = outs[j];
       int64_t out_part = parts[out];
@@ -801,11 +801,11 @@ while(t < edge_outer_iter)
       for (int64_t p = 0; p < num_parts; ++p)
         part_counts[p] = 0.0;
 
-      unsigned out_degree = out_degree(g, v);
+      uint64_t out_degree = out_degree(g, v);
       int64_t* outs = out_vertices(g, v);
       int64_t* weights = out_weights(g, v);
       int64_t sum_weights = 0;
-      for (unsigned j = 0; j < out_degree; ++j)
+      for (uint64_t j = 0; j < out_degree; ++j)
       {
         int64_t out = outs[j];
         int64_t part_out = parts[out];
@@ -873,7 +873,7 @@ while(t < edge_outer_iter)
             thread_queue_size = 0;
           }
         }
-        for (unsigned j = 0; j < out_degree; ++j)
+        for (uint64_t j = 0; j < out_degree; ++j)
         {
           if (!in_queue_next[outs[j]])
           {
@@ -1011,11 +1011,11 @@ while(t < edge_outer_iter)
       for (int64_t p = 0; p < num_parts; ++p)
         part_counts[p] = 0;
 
-      unsigned out_degree = out_degree(g, v);
+      uint64_t out_degree = out_degree(g, v);
       int64_t* outs = out_vertices(g, v);
       int64_t* weights = out_weights(g, v);
       int64_t sum_weights = 0;
-      for (unsigned j = 0; j < out_degree; ++j)
+      for (uint64_t j = 0; j < out_degree; ++j)
       {
         int64_t out = outs[j];
         int64_t part_out = parts[out];
@@ -1083,7 +1083,7 @@ while(t < edge_outer_iter)
               thread_queue_size = 0;
             }
           }
-          for (unsigned j = 0; j < out_degree; ++j) 
+          for (uint64_t j = 0; j < out_degree; ++j) 
           {
             if (!in_queue_next[outs[j]])
             {
